requirements.txt
fastapi
uvicorn
langchain
langchain-openai
youtube-transcript-api
faiss-cpu
python-dotenv
langchain-google-genai
